requirements.txt
anyio==3.5.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
contextlib2==21.6.0
contextvars==2.4
dataclasses==0.8
dnspython==2.2.0
email-validator==1.1.3
fastapi==0.74.0
h11==0.13.0
httptools==0.2.0
idna==3.3
immutables==0.16
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
orjson==3.6.1
pkg_resources==0.0.0
pydantic==1.9.0
PyJWT==2.3.0
python-decouple==3.6
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.14.0
watchgod==0.7
websockets==9.1
zipp==3.6.0